Create all the episodes views, create common components, update both reducers, rename components, optimization code
