+++ src/store/reducers/Characters.js
@@ -7,6 +7,7 @@ const initialState = {
 const charactersDataRecevied = (state, action) => {
 	const data = {
 		...action.data,
+		films: action.data.films.map((film)=>film.replace('https://swapi.co/api/films/', '')[0]),
 		hairColor: action.data.hair_color,
 		skinColor: action.data.skin_color,
 		eyeColor: action.data.eye_color
